Yet another performance improvement for Go 1.21+. Reuse maps backing the sliding windows to avoid map re-allocation, resizing and garbage collection when shifting the windows every second (or minute).

Speeds up throughput by 2.67x and reduces memory allocations by 2x on Go 1.21+ 🚀 

```
$ benchstat old.txt new.txt
goos: darwin
goarch: arm64
pkg: github.com/go-chi/httprate
               │   old.txt   │               new.txt               │
               │   sec/op    │   sec/op     vs base                │
LocalCounter-8   34.65m ± 3%   12.99m ± 2%  -62.52% (p=0.000 n=10)

               │   old.txt    │               new.txt                │
               │     B/op     │     B/op      vs base                │
LocalCounter-8   4.384Mi ± 0%   2.846Mi ± 0%  -35.07% (p=0.000 n=10)

               │   old.txt   │               new.txt               │
               │  allocs/op  │  allocs/op   vs base                │
LocalCounter-8   253.9k ± 0%   121.6k ± 0%  -52.12% (p=0.000 n=10)
```

(The CI runs Go 1.17, so we won't see any difference there.)